Change default value of 'async' for main Thread
Previously Propono.publish defaulted to 'async' mode unless a specific override
was made in the configuration hash. This is a great default for a webserver,
where publishing messages out-of-band is really important, but a frustrating
default when using Propono from a single-thread app. In that scenario, the main
application thread often exits before the messages have been sent

A better solution would probably to have a countdown-latch or similar guard on
Propono, but this helps make things cleaner for using Propono in a Rake task.